Move LoadingScreen out of 'mobile' Add LoadingScreen to desktop version of RootScreen
